requirements.txt
Flask
Flask-Caching
Flask-Bootstrap
Flask-Login
Flask-RESTful
Flask-Security-Too
Flask-SQLAlchemy
Flask-Mail
Flask-Migrate
flask-marshmallow
Flask-WTF
Werkzeug
python-dotenv
djlint
marshmallow_sqlalchemy
requests
celery
Flask-Excel
redis